requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Web Application
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0

# Data Augmentation
albumentations>=1.0.0
imgaug>=0.4.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=6.0.0
pytest-cov>=2.0.0

# Visualization
plotly>=5.0.0

# Model Deployment (optional)
# onnx>=1.10.0
# onnxruntime>=1.9.0
